+++ CHANGES.md
@@ -1,4 +1,8 @@
 #1.2.5
+ - Update to libGDX 1.9.12.
+ - **[Breaking]** Added `LiveWallpaperListener.setLiveWallpaperApplicationDelegate()`.
+ - Added LiveWallpaperApplicationDelegate for accessing `AndroidLiveWallpaper.notifyColorsChanged()`
+ from the core module.
 
 # 1.2.4
  - Update Android target to SDK 30.
@@ -17,7 +21,7 @@
  occur when different flavors of a project are installed on the same device.
 
 # 1.2.3
- - Update to LibGDX 1.9.11
+ - Update to libGDX 1.9.11.
  - Deprecated `Anisotropy`, as its features are now available from Texture directly.
  - **[Breaking (imports only)]** WallpaperEventListener is no longer an inner interface of LiveWallpaperWrapper.
  - Added WallpaperEventAdapter, empty implementation of WallpaperEventListener.
@@ -31,7 +35,7 @@
  - Add TextureAtlasCacher.
 
 # 1.2.1 
- - Update to LibGDX 1.9.10
+ - Update to libGDX 1.9.10.
  - **[Breaking]** `ColorUtil.blendRGBWithSaturation(...)` changed signature to include result parameter to avoid ambiguity
  - **[Breaking]** ColorUtil methods removed if they duplicated functionality from the Color class.
  - **[Breaking]** ColorUtil methods involving hue changed from 0-1 scale to 0-360 scale.